+++ frontend/src/components/search/Search.tsx
@@ -8,7 +8,7 @@ import useDebounceValue from '../../hooks/useDebouncedValue';
 import MaginifyingGlassIcon from '../../icons/MagnifyingGlassIcon';
 import { createDetailQuery } from '../../utils/createQuery';
 import { getDoiKey, isDoiFormat } from '../../utils/format';
-import { getLocalStorage, setLocalStorage } from '../../utils/localStorage';
+import { getLocalStorage, setLocalStorage } from '../../utils/storage';
 import IconButton from '../IconButton';
 import MoonLoader from '../loader/MoonLoader';
 import AutoCompletedList from './AutoCompletedList';
