+++ frontend/src/utils/storage.ts
@@ -1,3 +1,13 @@
+export const setLocalStorage = (key: string, value: unknown) => {
+  window.localStorage.setItem(key, JSON.stringify(value));
+};
+
+export const getLocalStorage = (key: string) => {
+  const item = window.localStorage.getItem(key);
+  if (item) return JSON.parse(item);
+  return null;
+};
+
 export const setSessionStorage = (key: string, value: unknown) => {
   window.sessionStorage.setItem(key, JSON.stringify(value));
 };
